Addresses PR Feedback
