--- pipelines/controller/src/main/java/com/google/fhir/analytics/DwhDto.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2020-2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.fhir.analytics;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
-import lombok.Data;
-
-@Data
-public class DwhDto {
-  @JsonProperty("dwh_prefix")
-  private String dwhPrefix;
-
-  @JsonProperty("dwh_path_latest")
-  private String dwhPath;
-
-  @JsonProperty("dwh_snapshot_ids")
-  private List<String> dwhSnapshots;
-}
